content_creation_agency/requirements.txt
agency-swarm==0.1.0
openai>=1.3.0
python-dotenv>=0.19.0
tavily-python>=0.1.9
nltk>=3.8.1
pytrends>=4.9.0
google-api-python-client>=2.108.0
textblob>=0.17.1
pandas>=2.1.3 